requirements.txt
annotated-types==0.7.0
certifi==2026.2.25
charset-normalizer==3.4.4
idna==3.11
iniconfig==2.3.0
markdown-it-py==4.0.0
mdurl==0.1.2
numpy==2.4.2
packaging==26.0
pandas==3.0.1
pluggy==1.6.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
requests==2.32.5
rich==14.3.3
six==1.17.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
